# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PublicLibrary API",
    "version": "1.0.0",
    "description": "API for PublicLibrary domain"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/patron/registerpatron": {
      "post": {
        "summary": "Execute RegisterPatron command",
        "operationId": "RegisterPatron",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPatronRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterPatronResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/patron/{patronId}/suspendpatron": {
      "post": {
        "summary": "Execute SuspendPatron command",
        "operationId": "SuspendPatron",
        "parameters": [
          {
            "name": "patronId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendPatronRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspendPatronResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/patron/{patronId}/restorepatron": {
      "post": {
        "summary": "Execute RestorePatron command",
        "operationId": "RestorePatron",
        "parameters": [
          {
            "name": "patronId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestorePatronRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestorePatronResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/patron/{patronId}": {
      "get": {
        "summary": "Get Patron by ID",
        "operationId": "getPatronById",
        "parameters": [
          {
            "name": "patronId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patron found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patron"
                }
              }
            }
          },
          "404": {
            "description": "Patron not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Patron": {
        "type": "object",
        "properties": {
          "patronId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "researcherStatus": {
            "type": "boolean"
          },
          "unpaidFeeBalance": {
            "type": "number",
            "format": "decimal"
          }
        },
        "required": [
          "patronId",
          "name",
          "researcherStatus",
          "unpaidFeeBalance"
        ]
      },
...
      "RegisterPatronRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "researcherStatus": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "researcherStatus"
        ]
      },
      "RegisterPatronResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "patronId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "SuspendPatronRequest": {
        "type": "object",
        "properties": {
          "patronId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "patronId",
          "patronId"
        ]
      },
      "SuspendPatronResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "patronId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "RestorePatronRequest": {
        "type": "object",
        "properties": {
          "patronId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "patronId",
          "patronId"
        ]
      },
      "RestorePatronResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "patronId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}